debug(auth): add logging to trace encryption key through login flow
Added debug logging to see what sessionData contains when passed to setAuthenticated.

This will help us understand why encryptionKey is not making it to
the auth context even though we're passing result.sessionData.

Once deployed, logs will show:
- hasSessionData: should be true
- hasEncryptionKey: should be true
- encryptionKeyType: should be 'CryptoKey'
- hasSalt: should be true
- saltType: should be 'Uint8Array'

If any of these are false/undefined, we'll know where the problem is.